@lemire 2018-08-09
This looks like a sane optimization.